Ignite.jl

Documentation for Ignite.jl.

Docstrings

Ignite.Ignite

Ignite.AbstractEvent

Ignite.AbstractPrimitiveErrorEvent

Ignite.AbstractPrimitiveEvent

Ignite.AndEvent

Ignite.Engine

Ignite.EventHandler

Ignite.FilteredEvent

Ignite.OrEvent

Ignite.State

Ignite.add_event_handler!

Ignite.every_filter

Ignite.fire_event!

Ignite.once_filter

Ignite.run!

Ignite.@on

Ignite.Ignite — Module

Ignite.jl

<!– –>

Welcome to Ignite.jl, a Julia port of the Python library ignite for simplifying neural network training and validation loops using events and handlers.

Ignite.jl provides a simple yet flexible engine and event system, allowing for the easy composition of training pipelines with various events such as artifact saving, metric logging, and model validation. Event-based training abstracts away the training loop, replacing it with 1) a training engine which executes a single training step, 2) a data loader for the engine to iterate over, and 3) events and corresponding handlers which are attached to the engine, configured to fire at specific points during training.

Event handlers much more flexibile compared to other approaches like callbacks: they can be any callable, multiple handlers can be attached to a single event, multiple events can trigger the same handler, and custom events can be defined to fire at user-specified points during training. This makes adding functionality to your training pipeline easy, minimizing the need to modify existing code.

Quick Start

The example below demonstrates how to use Ignite.jl to train a simple neural network while keeping track of evaluation metrics and displaying them every 5 epochs. Key features to note:

The training step is factored out of the training loop: the train_step function takes a batch of training data and computes the loss, gradients, and updates the model parameters.

Events are flexibly added to the training and evaluation engines to customize training: the evaluator logs metrics, and the trainer runs the evaluator every 5 epochs.

Data loaders can be any iterable collection. Here we use a DataLoader from MLUtils.jl

using Ignite
using Flux, Zygote, Optimisers, MLUtils # for training a neural network
using OnlineStats: Mean, fit! # for tracking evaluation metrics

# Build simple neural network and initialize Adam optimizer
model = Chain(Dense(1 => 32, tanh), Dense(32 => 1))
optim = Optimisers.setup(Optimisers.Adam(1f-3), model)

# Creat mock data and data loaders
f(x) = 2x-x^3
xtrain, xtest = randn(1, 10_000), randn(1, 100)
ytrain, ytest = f.(xtrain), f.(xtest)
train_data_loader = DataLoader((; x = xtrain, y = ytrain); batchsize = 64, shuffle = true, partial = false)
eval_data_loader = DataLoader((; x = xtest, y = ytest); batchsize = 10, shuffle = false)

# Create training engine:
#   - `engine` is a reference to the parent `trainer` engine, created below
#   - `batch` is a batch of training data, retrieved by iterating `train_data_loader`
#   - (optional) return value is stored in `trainer.state.output`
function train_step(engine, batch)
    x, y = batch
    l, gs = Zygote.withgradient(m -> sum(abs2, m(x) .- y), model)
    global optim, model = Optimisers.update!(optim, model, gs[1])
    return Dict("loss" => l)
end
trainer = Engine(train_step)

# Create evaluation engine using `do` syntax:
evaluator = Engine() do engine, batch
    x, y = batch
    ypred = model(x) # evaluate model on a single batch of validation data
    return Dict("ytrue" => y, "ypred" => ypred) # result is stored in `evaluator.state.output`
end

# Add events to the evaluation engine to track metrics:
#   - when `evaluator` starts, initialize the running mean
#   - after each iteration, compute eval metrics from predictions and update the running average
add_event_handler!(evaluator, STARTED()) do engine
    engine.state.metrics = Dict("abs_err" => Mean()) # new fields can be dynamically added to `engine.state`
end
add_event_handler!(evaluator, ITERATION_COMPLETED()) do engine
    o = engine.state.output
    m = engine.state.metrics["abs_err"]
    fit!(m, abs.(o["ytrue"] .- o["ypred"]) |> vec)
end

# Add an event to the training engine to run `evaluator` every 5 epochs:
add_event_handler!(trainer, EPOCH_COMPLETED(every = 5)) do engine
    Ignite.run!(evaluator, eval_data_loader)
    @info "Evaluation metrics: abs_err = $(evaluator.state.metrics["abs_err"])"
end

# Start the training
Ignite.run!(trainer, train_data_loader; max_epochs = 25, epoch_length = 1_000)

Periodically save model

Easily add custom functionality to your training process without modifying existing code by incorporating new events. For example, saving the current model and optimizer state to disk every 10 epochs using BSON.jl:

using BSON

# Save model and optimizer state every 10 epochs
add_event_handler!(trainer, EPOCH_COMPLETED(every = 10)) do engine
    @save "model_and_optim.bson" model optim
    @info "Saved model and optimizer state to disk"
end

Execute any number of functions per event

Multiple handlers can be added to the same event:

function on_training_ended(engine)
    @info "Training is ended"
end
add_event_handler!(on_training_ended, trainer, COMPLETED())
add_event_handler!(engine -> display(engine.state.times), trainer, COMPLETED())

Attach the same handler to multiple events

The boolean operators | and & can be used to combine events:

add_event_handler!(trainer, COMPLETED() | EPOCH_COMPLETED(every = 10)) do engine
    # ...
end

Fire custom events

Custom events can be created to track different stages in the training process, such as the start and finish of the backward pass and optimizer step, by defining new event types that inherit from AbstractPrimitiveEvent and firing them at appropriate points in the trainstep function using `fireevent!`:

struct BACKWARD_STARTED <: AbstractPrimitiveEvent end
struct BACKWARD_COMPLETED <: AbstractPrimitiveEvent end
struct OPTIM_STEP_STARTED <: AbstractPrimitiveEvent end
struct OPTIM_STEP_COMPLETED <: AbstractPrimitiveEvent end

function train_step(engine, batch)
    x, y = batch

    # Compute the gradients of the loss with respect to the model
    fire_event!(engine, BACKWARD_STARTED())
    l, gs = Zygote.withgradient(m -> sum(abs2, m(x) .- y), model)
    fire_event!(engine, BACKWARD_COMPLETED())

    # Update the model's parameters
    fire_event!(engine, OPTIM_STEP_STARTED())
    global optim, model = Optimisers.update!(optim, model, gs[1])
    fire_event!(engine, OPTIM_STEP_COMPLETED())

    return Dict("loss" => l)
end
trainer = Engine(train_step)

Ignite.AbstractEvent — Type

Abstract event type for construction compound events.

Ignite.AbstractPrimitiveErrorEvent — Type

Basic events triggered by errors.

Ignite.AbstractPrimitiveEvent — Type

Basic events fired by the engine.

Ignite.AndEvent — Type

struct AndEvent{E1<:AbstractEvent, E2<:AbstractEvent} <: AbstractEvent

AndEvent(event1::E1, event2::E2) wraps two events and triggers if and only if both wrapped events are triggered by the same primitive event firing.

Fields:

event1::AbstractEvent: The first wrapped event that will be considered for triggering.

event2::AbstractEvent: The second wrapped event that will be considered for triggering.

Ignite.Engine — Type

mutable struct Engine{P}

Engine(process_function::P; kwargs...) to be run; see Ignite.run!.

Fields:

process_function::Any: A function that processes a single batch of data and returns an output.

dataloader::Any: An iterable object that provides the data for the engine to process. Default: nothing

state::State: An object that holds the current state of the engine. Default: State()

event_handlers::Vector{EventHandler}: A list of event handlers that are called at specific points when the engine is running. Default: EventHandler[]

logger::Union{Nothing, Base.CoreLogging.AbstractLogger}: An optional logger; if nothing, then current_logger() will be used. Default: nothing

should_terminate::Bool: A flag that indicates whether the engine should stop running. Default: false

_dataloader_iter::Any: Default: nothing

Ignite.EventHandler — Type

struct EventHandler{E<:AbstractEvent, H}

EventHandler(event::E, handler!::H = Returns(nothing)) wraps an event and fires the corresponding handler! when event is triggered.

Fields:

event::AbstractEvent: Event which triggers handler

handler!::Any: Event handler which executes when triggered by event Default: Returns(nothing)

Ignite.FilteredEvent — Type

struct FilteredEvent{E<:AbstractEvent, F} <: AbstractEvent

FilteredEvent(event::E, filter::F = Returns(true)) wraps an event and a filter function.

When a primitive event e is fired, if filter(engine, e) returns true then the filtered event will be fired too.

Fields:

event::AbstractEvent: The wrapped event that will be fired if the filter function returns true when applied to a primitive event.

filter::Any: The filter function filter(engine::Engine, e::AbstractPrimitiveEvent)::Bool returns true if the wrapped event should be fired. Default: Returns(true)

Ignite.OrEvent — Type

struct OrEvent{E1<:AbstractEvent, E2<:AbstractEvent} <: AbstractEvent

OrEvent(event1::E1, event2::E2) wraps two events and triggers if either of the wrapped events are triggered by a primitive event firing.

Fields:

event1::AbstractEvent: The first wrapped event that will be checked if it should be fired.

event2::AbstractEvent: The second wrapped event that will be checked if it should be fired.

Ignite.State — Type

struct State

Current state of the engine. Fields can be accessed and modified using getproperty and setproperty!.

State is a light wrapper around a DefaultOrderedDict{Symbol, Any, Nothing} with the following keys:

:iteration: the current iteration, beginning with 1.

:epoch: the current epoch, beginning with 1.

:dataloader: The data loader passed to the engine.

:max_epochs: The number of epochs to run.

:epoch_length: The number of batches processed per epoch.

:batch: The current batch passed to process_function.

:output: The output of process_function after a single iteration.

:last_event: The last event fired.

:counters: A DefaultOrderedDict{AbstractPrimitiveEvent, Int, Int}(0) with primitive event firing counters.

:times: An OrderedDict{AbstractPrimitiveEvent, Float64}() with total and per-epoch times fetched on primitive event keys.

For example, engine.state.iteration can be used to access the current iteration.

Ignite.add_event_handler! — Method

add_event_handler!(
    handler,
    engine::Engine,
    event::AbstractEvent
) -> Engine

Add event handler to engine which is fired when event is triggered.

Ignite.every_filter — Method

every_filter(; every)

Creates a filter function for use in a FilteredEvent that returns true periodically depending on every:

If every = n::Int, the filter will trigger every nth firing of the event.

If every = Int[n₁, n₂, ...], the filter will trigger every n₁th firing, every n₂th firing, and so on.

Ignite.fire_event! — Method

fire_event!(
    engine::Engine,
    handler::EventHandler,
    e::AbstractPrimitiveEvent
) -> Engine

Execute handler if it is triggered by the primitive event e.

Ignite.once_filter — Method

once_filter(; once)

Creates a filter function for use in a FilteredEvent that returns true at specific points depending on once:

If once = n::Int, the filter will trigger only on the nth firing of the event.

If once = Int[n₁, n₂, ...], the filter will trigger only on the n₁th firing, the n₂th firing, and so on.

Ignite.run! — Function

run!(engine::Engine) -> Engine
run!(
    engine::Engine,
    dataloader;
    max_epochs,
    epoch_length
) -> Engine

Run the engine. Data batches are retrieved by iterating dataloader. The data loader may be infinite; by default, it is wrapped in Iterators.cycle to restart if it empties.

Inputs:

engine::Engine: An instance of the Engine struct containing the process_function to run each iteration.

dataloader: A data loader to iterate over. Defaults to engine.dataloader.

max_epochs::Int: the number of epochs to run. Defaults to 1.

epoch_length::Int: the length of an epoch. If not given, will try to default to length(dataloader).

Conceptually, running the engine is roughly equivalent to the following:

The engine state is initialized.

The engine begins running for max_epochs epochs, or until the should_terminate flag is set to true.

At the start of each epoch, EPOCH_STARTED() event is fired and time is recorded.

An iteration loop is performed for epoch_length number of iterations, or until the should_terminate flag is set to true.

At the start of each iteration, ITERATION_STARTED() event is fired, and a batch of data is loaded.

The process_function is called on the loaded data batch.

At the end of each iteration, ITERATION_COMPLETED() event is fired.

At the end of each epoch, EPOCH_COMPLETED() event is fired.

At the end of all the epochs, COMPLETED() event is fired.

Finally, TERMINATE() event is fired if should_terminate flag is set to true.

Ignite.@on — Macro

@on engine event handler is syntax sugar for add_event_handler!(handler, engine, event).